Revert "Update OMSDK to version 1.3.12"
This reverts commit 74315014dbcf4e4e4f6b71ada1163ca878ae96e4.
