+++ OMSDK_Verizonmedia.framework/Versions/A/Headers/OMIDImports.h
@@ -6,4 +6,4 @@
 #import "OMIDAdSession.h"
 #import "OMIDAdEvents.h"
 #import "OMIDVASTProperties.h"
-#import "OMIDMediaEvents.h"
+#import "OMIDVideoEvents.h"
